updated gamepad configuration ui
buttons now change color to indicate when they're looking for joystick input and when they're calibrated/set.
The floating point numbers in the app now display "0.000 0.000" until the gamepad has been configured. When the
gamepad is set it will then display joystick axes position for the left and right motor